fix addEffectsCount
added effectsCount.. which meant I had to touch all the effects and
remove their particular Count vars
Also put reset/setup of timer/count into SetNoEfx() and use it in start
of all Set…Efx().
While doing this I found the RainbowEfx subclass was a bit lacking.  It
only implemented one of three types in strand test example… The evenly
distributed version was not implemented, but useful.  So I put in two
RainbowTypes.
And then implemented a state machine for the two rings/windows in the
Subclass example app.